pyproject.toml
[tool.poetry]
name = "dharan-backend"
version = "0.1.0"
description = ""
authors = ["Arpan Koirala <dazehere@yandex.com>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.70.0"
uvicorn = {version = "^0.15.0", extras = ["standard"]}
gunicorn = "^20.1.0"
pydantic = "^1.8.2"
SQLAlchemy = "^1.4.25"
psycopg2-binary = "^2.9.1"
alembic = "^1.7.4"
passlib = {version = "^1.7.4", extras = ["argon2"]}
cryptography = "^35.0.0"
aioredis = "^2.0.0"
python-dotenv = "^0.19.1"
click = "^8.0.3"
pyotp = "^2.6.0"
email-validator = "^1.1.3"
python-multipart = "^0.0.5"
requests = "^2.26.0"
Authlib = "^0.15.5"
emails = "^0.6"
Jinja2 = "^3.0.2"
python-jose = "^3.3.0"

[tool.poetry.dev-dependencies]
jedi = "^0.18.0"
rope = "^0.20.1"
pylint = "^2.11.1"
black = "^21.9b0"
mypy = "^0.910"
pytest = "^6.2.5"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
